+++ src/expression_tree/object.jl
@@ -1,5 +1,5 @@
 """
-    mutable struct Propagator{PARA,F}
+    mutable struct Propagator{PARA, F, N}
 
      Object of all kinds of propagators (many-body Green's functions, interactions, vertex functions, ...)
 
@@ -19,7 +19,7 @@ struct Propagator{PARA,F}
     # function Propagator(order, basis = [], factor::F = 1.0, para::P = 0) where {F,P}
     #     return new{P,F}(para, order, factor, basis)
     # end
-    function Propagator{F}(name::Symbol, order::Int, para::P, factor, loopidx::Int, sitebasis) where {F,P}
+    function Propagator{P,F}(name::Symbol, order::Int, para, factor, loopidx::Int, sitebasis) where {P,F}
         return new{P,F}(name, para, order, F(factor), loopidx, sitebasis)
     end
 end
@@ -57,7 +57,7 @@ struct Node{PARA,F}
     operation::Int #1: multiply, 2: add, ...
     factor::F
     # order::Int
-    propagators::Vector{Vector{Int}}
+    propagators::Vector{Int}
     childNodes::Vector{Int}
     parent::Int # parent id
     # child::SubArray{CachedObject{Node{PARA, P},W},1,Vector{CachedObject{NODE,W}},Tuple{Vector{Int64}},false}
@@ -65,11 +65,11 @@ struct Node{PARA,F}
     # function Node(operation::Int, components = [[]], child = [], factor::F = 1.0, parent = 0, para::P = :N) where {F,P}
     #     return new{P,F}(para, operation, factor, components, child, parent)
     # end
-    function Node{F}(name::Symbol, operation::Int, para::P, propagators = [[]], child = [], factor = 1.0, parent = 0) where {F,P}
+    function Node{F}(name::Symbol, operation::Int, para::P, propagators = [], child = [], factor = 1.0, parent = 0) where {F,P}
         # @assert typeof(para) == P
         return new{P,F}(name, para, operation, F(factor), propagators, child, parent)
     end
-    function Node{P,F}(name::Symbol, operation::Int, para, propagators = [[]], child = [], factor = 1.0, parent = 0) where {F,P}
+    function Node{P,F}(name::Symbol, operation::Int, para, propagators = [], child = [], factor = 1.0, parent = 0) where {F,P}
         # @assert typeof(para) == P
         return new{P,F}(name, para, operation, F(factor), propagators, child, parent)
     end
@@ -119,20 +119,14 @@ mutable struct Diagrams{V,P,PARA,F,W}
     #     return new{V,P,PARA,F,W}(basisPool, propagatorPool, Pool{Cache{Node{PARA,F},W}}(), [])
     # end
     function Diagrams(basisPool::V, propagatorPool::P, nodeWeightType::DataType; nodeFactorType = nodeWeightType, nodeParaType::DataType = Nothing, name = :none) where {V,P}
-        # typeof(nothing) = Nothing
-        # @assert V <: Tuple "Tuple is required for efficiency!"
-        @assert P <: Tuple "Tuple is required for efficiency!"
-        # println(basisPool)
-        # println(propagatorPool)
-        poolname = []
-        for p in propagatorPool
-            push!(poolname, p.name)
-        end
-        @assert length(poolname) == length(Set(poolname)) "All propagatorPool should have different names, yet got $poolname!"
         nodePool = CachedPool(:node, Node{nodeParaType,nodeFactorType}, nodeWeightType)
         return new{V,P,nodeParaType,nodeFactorType,nodeWeightType}(name, basisPool, propagatorPool, nodePool, [])
     end
 end
+
+Base.getindex(diag::Diagrams, i) = diag.nodePool.current[diag.root[i]]
+Base.firstindex(diag::Diagrams) = 1
+Base.lastindex(diag::Diagrams) = length(diag.root)
 
 struct Component
     index::Int
@@ -172,19 +166,19 @@ end
 - loop = nothing : loop basis (e.g, momentum and frequency) of the propagator.
 - para = nothing : Additional paramenter required to evaluate the propagator.
 """
-function addPropagator!(diag::Diagrams, index::Int, order::Int, name, factor = 1.0; site = [], loop = nothing, para = nothing)
+function addPropagator!(diag::Diagrams, name, factor = 1.0; site = [], loop = nothing, para = nothing, order::Int = 0)
     loopPool = diag.basisPool
     propagatorPool = diag.propagatorPool
     # @assert length(basis) == length(variablePool) == length(currVar) "$(length(basis)) == $(length(variablePool)) == $(length(currVar)) breaks"
 
-    @assert 0 < index <= length(propagatorPool) "proapgator Pool index $index is illegal!"
+    # @assert 0 < index <= length(propagatorPool) "proapgator Pool index $index is illegal!"
 
-    PROPAGATOR_POOL = typeof(propagatorPool[index])
+    PROPAGATOR_POOL = typeof(propagatorPool)
     PROPAGATOR = eltype(fieldtype(PROPAGATOR_POOL, :object))
     PARA = fieldtype(PROPAGATOR, :para)
     F = fieldtype(PROPAGATOR, :factor)
 
-    @assert PARA == typeof(para) "Type of $para is not $PARA"
+    @assert typeof(para) <: PARA "Type of $para is $(typeof(para)), while we expect $PARA"
 
     # function Propagator{P,F}(order, para, factor, loopbasis, localbasis) where {P,F}
     loopidx = 0
@@ -192,33 +186,10 @@ function addPropagator!(diag::Diagrams, index::Int, order::Int, name, factor = 1
         @assert typeof(loop) <: AbstractVector "LoopBasis should be a Vector!"
         loopidx = append(loopPool, loop)
     end
-    prop = Propagator{F}(name, order, para, factor, loopidx, collect(site))
-    pidx = append(diag.propagatorPool[index], prop)
+    prop = Propagator{PARA,F}(name, order, para, factor, loopidx, collect(site))
+    pidx = append(diag.propagatorPool, prop)
     # return component(pidx, false, propagatorPool[index].name)
     return pidx
-end
-"""
-    function addPropagator!(diag::Diagrams, poolName::Symbol, order::Int, name, factor = 1; site = [], loop = nothing, para = nothing)
-
-    Add a propagator into the diagram tree referenced by the pool name.
-
-# Arguments
-- diag::Diagrams : Diagram tree.
-- poolName::Symbol : name of the propagator pool
-- order::Int     : Order of the propagator.
-- name = :none   : name of the propagator.
-- factor = 1     : Factor of the propagator.
-- site = []      : site basis (e.g, time and space coordinate) of the propagator.
-- loop = nothing : loop basis (e.g, momentum and frequency) of the propagator.
-- para = nothing : Additional paramenter required to evaluate the propagator.
-"""
-function addPropagator!(diag::Diagrams, poolName::Symbol, order::Int, name, factor = 1.0; site = [], loop = nothing, para = nothing)
-    for (idx, pool) in enumerate(diag.propagatorPool)
-        if pool.name == poolName
-            return addPropagator!(diag, idx, order, name, factor; site = site, loop = loop, para = para)
-        end
-    end
-    error("Propagator poolName $poolName doesn't exist in $([pool.name for pool in diag.propagatorPool])!")
 end
 
 """
@@ -239,15 +210,15 @@ function addNode!(diag::Diagrams, operator, name, factor = 1.0; propagator = not
     nodePool = diag.nodePool
 
     if isnothing(propagator)
-        propagator = [[] for p in diag.propagatorPool]
+        propagator = []
     end
 
-    @assert length(propagator) == length(diag.propagatorPool) "each element of the propagator is an index vector of the corresponding propagator"
+    # @assert length(propagator) == length(diag.propagatorPool) "each element of the propagator is an index vector of the corresponding propagator"
 
     filterZero(list) = [l for l in list if l != 0]
 
     #filter the propagators and nodes with index 0, they are the empty object
-    propagator = [filterZero(c) for c in propagator]
+    propagator = filterZero(propagator)
     childNodes = filterZero(child)
 
     empty = true
@@ -273,10 +244,8 @@ function addNode!(diag::Diagrams, operator, name, factor = 1.0; propagator = not
     # println("node F: ", F)
     # @assert PARA == typeof(para) "Type of $para is not $PARA"
 
-    for (i, p) in enumerate(propagator)
-        for pidx in p
-            @assert pidx <= length(diag.propagatorPool[i]) "Failed to add node with propagator = $propagator, and child =$childNodes. $pidx is not in pool $i (length = $(length(diag.propagatorPool[i])))."
-        end
+    for pidx in propagator
+        @assert pidx <= length(diag.propagatorPool) "Failed to add node with propagator = $propagator, and child =$childNodes. $pidx is not in GW pool (length = $(length(diag.propagatorPool)))."
     end
     for nidx in childNodes
         @assert nidx <= length(diag.nodePool) "Failed to add node with propagator = $propagator, and child =$childNodes. $nidx is not in nodePool."
@@ -288,32 +257,6 @@ function addNode!(diag::Diagrams, operator, name, factor = 1.0; propagator = not
     nidx = append(nodePool, node)
     return nidx
     # return component(nidx, true, :none)
-end
-"""
-    function addNodeByName(diag::Diagrams, operator, name = :none; child = [], factor = 1.0, parent = 0, para = nothing, kwargs...)
-
-    An alternative way to add node. Instead of providing the components vector, one specify the propagators in kwargs. 
-
-# Arguments
-- diag::Diagrams : Diagram tree.
-- operator::Int  : #1: multiply, 2: add, ...
-- name           : name of the node
-- factor = 1.0   : Factor of the node
-- child = []        : Indices to the cached nodes stored in certain pool. They are the child of the current node in the diagram tree. It should be in the format of Vector{Int}.
-- para = nothing    : Additional paramenter required to evaluate the node. Set to nothing by default.
-- kwargs...      : Accept arbitrary pair of parameters as: PropagatorPoolName = [index of the propagator1, index of the propagator2, ...]
-"""
-function addNodeByName!(diag::Diagrams, operator, name, factor = 1.0; child = [], para = nothing, kwargs...)
-    components = [[] for p in diag.propagatorPool]
-    for key in keys(kwargs)
-        for (idx, p) in enumerate(diag.propagatorPool)
-            if p.name == key
-                append!(components[idx], kwargs[key])
-            end
-        end
-    end
-    # println(kwargs, " got ", components)
-    return addNode!(diag, operator, name, factor; propagator = components, child = child, para = para)
 end
 
 """
@@ -329,14 +272,14 @@ function getNode(diag::Diagrams, n::Component)
     return diag.nodePool.object[n.index]
 end
 
-function getPropagator(diag::Diagrams, pidx::Int, poolName::Symbol)
-    for p in diag.propagatorPool
-        if p.name == poolName
-            return p.object[pidx]
-        end
-    end
-    error("$poolName propagator pool doesn't exist!")
-end
+# function getPropagator(diag::Diagrams, pidx::Int)
+#     for p in diag.propagatorPool
+#         if p.name == poolName
+#             return p.object[pidx]
+#         end
+#     end
+#     error("$poolName propagator pool doesn't exist!")
+# end
 
 """
     function getNodeWeight(diag::Diagrams, nidx::Int)
@@ -348,10 +291,10 @@ function getNodeWeight(diag::Diagrams, nidx::Int)
 end
 
 
-function addpropagator!(diag::Diagrams, poolName::Symbol, order::Int, name, factor = 1.0; site = [], loop = nothing, para = nothing)
-    pidx = addPropagator!(diag, poolName, order, name, factor; site = site, loop = loop, para = para)
+function addpropagator!(diag::Diagrams, name, factor = 1.0; site = [], loop = nothing, para = nothing, order::Int = 0)
+    pidx = addPropagator!(diag, name, factor; site = site, loop = loop, para = para, order = order)
     @assert pidx > 0
-    return Component(pidx, false, poolName, getPropagator(diag, pidx, poolName))
+    return Component(pidx, false, :propagator, diag.propagatorPool.object[pidx])
 end
 
 function addnode!(diag::Diagrams, operator, name, components, factor = 1.0; para = nothing)
@@ -367,38 +310,15 @@ function addnode!(diag::Diagrams, operator, name, components, factor = 1.0; para
     end
 
     child = []
-    propagator = [[] for p in diag.propagatorPool]
+    propagator = []
     for c in collect(_components)
         @assert c isa Component "$c is not a DiagTree.Component"
         if c.isNode
             push!(child, c.index)
         else
-            for (idx, p) in enumerate(diag.propagatorPool)
-                if p.name == c.poolName
-                    append!(propagator[idx], c.index)
-                end
-            end
+            push!(propagator, c.index)
         end
     end
     nidx = addNode!(diag, operator, name, factor; propagator = propagator, child = child, para = para)
     return Component(nidx, true, diag.nodePool.name, getNode(diag, nidx))
-end
-
-function sum_of_producted_components!(diag::Diagrams, name, _componentsVector::Vector{Vector{Component}}, factor = 1.0; para = nothing)
-    componentsVector = []
-    for components in _componentsVector
-        _components = [c for c in components if c.index > 0]
-        if isempty(_components)
-            @warn("Some of the components are emtpy! $componentsVector")
-        end
-        push!(componentsVector, _components)
-    end
-
-    nodes = []
-    for cs in Iterators.product(Tuple(componentsVector)...)
-        # println(typeof(cs))
-        nodename = reduce(*, [String(c.object.name) for c in cs])
-        push!(nodes, addnode!(diag, MUL, Symbol(nodename), cs; para = para))
-    end
-    return addnode!(diag, ADD, name, nodes; para = para)
 end
