--- src/expression_tree_backup/ast.jl
@@ -1,3 +1,0 @@
-function evalAST(idx, extT, loopidx)
-
-end
