--- src/expression_tree_backup/common.jl
@@ -1,64 +1,0 @@
-import ..Filter
-import ..Wirreducible  #remove all polarization subdiagrams
-import ..Girreducible  #remove all self-energy inseration
-import ..NoHatree
-import ..NoFock
-import ..NoBubble  # true to remove all bubble subdiagram
-import ..Proper  #ver4, ver3, and polarization diagrams may require to be irreducible along the transfer momentum/frequency
-
-import ..DiagramType
-import ..GreenDiag
-import ..SigmaDiag
-import ..PolarDiag
-import ..Ver3Diag
-import ..Ver4Diag
-
-import ..Composite
-import ..ChargeCharge
-import ..SpinSpin
-import ..UpUp
-import ..UpDown
-import ..Response
-
-import ..Instant
-import ..Dynamic
-import ..D_Instant
-import ..D_Dynamic
-import ..AnalyticProperty
-
-import ..symbol
-import ..short
-
-import ..Interaction
-import ..GenericPara
-import ..innerTauNum
-
-import ..Diagram
-
-import ..DiagramId
-import ..Ver4Id
-import ..Ver3Id
-import ..GreenId
-import ..SigmaId
-import ..PolarId
-import ..InteractionId
-
-import ..TwoBodyChannel
-import ..Alli
-import ..PHr
-import ..PHEr
-import ..PPr
-import ..AnyChan
-
-import ..Permutation
-import ..Di
-import ..Ex
-import ..DiEx
-
-import ..uidreset
-import ..toDataFrame
-import ..mergeby
-
-import ..Operator
-import ..Add
-import ..Prod
